backend/analytics/views.py
from rest_framework import status
from rest_framework.decorators import api_view, permission_classes
from rest_framework.response import Response
from rest_framework.permissions import IsAuthenticated
from django.utils import timezone
from datetime import datetime, timedelta, date


@api_view(['GET'])
@permission_classes([IsAuthenticated])
def dashboard_analytics(request):
    """Get dashboard analytics overview"""
    try:
        # Get current date and calculate date ranges
        today = timezone.now().date()
        last_week = today - timedelta(days=7)
        last_month = today - timedelta(days=30)
        
        # Basic analytics data
        analytics_data = {
            'total_metrics': 0,
            'total_insights': 0,
            'total_performance': 0,
            'date_ranges': {
                'today': today.isoformat(),
                'last_week': last_week.isoformat(),
                'last_month': last_month.isoformat(),
            }
        }
        
        return Response({
            'success': True,
            'data': analytics_data
        }, status=status.HTTP_200_OK)
        
    except Exception as e:
        return Response({
            'success': False,
            'message': 'Failed to get dashboard analytics.',
            'errors': {'detail': str(e)}
        }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)


@api_view(['GET'])
@permission_classes([IsAuthenticated])
def business_metrics(request):
    """Get business metrics list"""
    try:
        return Response({
            'success': True,
            'data': {
                'metrics': [],
                'pagination': {
                    'page': 1,
                    'page_size': 20,
                    'total_count': 0,
                    'total_pages': 1,
                    'has_next': False,
                    'has_previous': False
                }
            }
        }, status=status.HTTP_200_OK)
        
    except Exception as e:
        return Response({
            'success': False,
            'message': 'Failed to get business metrics.',
            'errors': {'detail': str(e)}
        }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)


@api_view(['GET'])
@permission_classes([IsAuthenticated])
def business_metric_detail(request, metric_id):
    """Get specific business metric detail"""
    try:
        return Response({
            'success': True,
            'data': {
                'id': str(metric_id),
                'message': 'Analytics module is being set up'
            }
        }, status=status.HTTP_200_OK)
        
    except Exception as e:
        return Response({
            'success': False,
            'message': 'Failed to get business metric detail.',
            'errors': {'detail': str(e)}
        }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)


@api_view(['GET'])
@permission_classes([IsAuthenticated])
def customer_insights(request):
    """Get customer insights list"""
    try:
        return Response({
            'success': True,
            'data': {
                'insights': [],
                'pagination': {
                    'page': 1,
                    'page_size': 20,
                    'total_count': 0,
                    'total_pages': 1,
                    'has_next': False,
                    'has_previous': False
                }
            }
        }, status=status.HTTP_200_OK)
        
    except Exception as e:
        return Response({
            'success': False,
            'message': 'Failed to get customer insights.',
            'errors': {'detail': str(e)}
        }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)


@api_view(['GET'])
@permission_classes([IsAuthenticated])
def customer_insight_detail(request, insight_id):
    """Get specific customer insight detail"""
    try:
        return Response({
            'success': True,
            'data': {
                'id': str(insight_id),
                'message': 'Analytics module is being set up'
            }
        }, status=status.HTTP_200_OK)
        
    except Exception as e:
        return Response({
            'success': False,
            'message': 'Failed to get customer insight detail.',
            'errors': {'detail': str(e)}
        }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)


@api_view(['GET'])
@permission_classes([IsAuthenticated])
def product_performance(request):
    """Get product performance list"""
    try:
        return Response({
            'success': True,
            'data': {
                'performance': [],
                'pagination': {
                    'page': 1,
                    'page_size': 20,
                    'total_count': 0,
                    'total_pages': 1,
                    'has_next': False,
                    'has_previous': False
                }
            }
        }, status=status.HTTP_200_OK)
        
    except Exception as e:
        return Response({
            'success': False,
            'message': 'Failed to get product performance.',
            'errors': {'detail': str(e)}
        }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)


@api_view(['GET'])
@permission_classes([IsAuthenticated])
def product_performance_detail(request, performance_id):
    """Get specific product performance detail"""
    try:
        return Response({
            'success': True,
            'data': {
                'id': str(performance_id),
                'message': 'Analytics module is being set up'
            }
        }, status=status.HTTP_200_OK)
        
    except Exception as e:
        return Response({
            'success': False,
            'message': 'Failed to get product performance detail.',
            'errors': {'detail': str(e)}
        }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)


@api_view(['GET'])
@permission_classes([IsAuthenticated])
def realtime_analytics(request):
    """Get real-time analytics data"""
    try:
        # Get current time-based analytics
        now = timezone.now()
        today = now.date()
        
        realtime_data = {
            'current_time': now.isoformat(),
            'today_date': today.isoformat(),
            'active_metrics': 0,
            'recent_insights': 0,
            'performance_updates': 0,
        }
        
        return Response({
            'success': True,
            'data': realtime_data
        }, status=status.HTTP_200_OK)
        
    except Exception as e:
        return Response({
            'success': False,
            'message': 'Failed to get real-time analytics.',
            'errors': {'detail': str(e)}
        }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)


@api_view(['GET'])
@permission_classes([IsAuthenticated])
def export_analytics(request):
    """Export analytics data"""
    try:
        # Get export parameters
        export_type = request.GET.get('type', 'all')
        date_from = request.GET.get('date_from')
        date_to = request.GET.get('date_to')
        
        export_data = {
            'export_type': export_type,
            'date_range': {
                'from': date_from,
                'to': date_to,
            },
            'timestamp': timezone.now().isoformat(),
            'message': 'Export functionality will be implemented based on requirements'
        }
        
        return Response({
            'success': True,
            'data': export_data
        }, status=status.HTTP_200_OK)
        
    except Exception as e:
        return Response({
            'success': False,
            'message': 'Failed to export analytics.',
            'errors': {'detail': str(e)}
        }, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
